pyproject.toml
[tool.poetry]
name = "highkick-agent"
version = "0.1.0"
description = ""
authors = ["Leonid <lganzzzo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
boto3 = "^1.34.128"
aioboto3 = "^13.1.1"
types-aiobotocore-s3 = "^2.15.2"
email-validator = "^2.2.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
